refactor(checkbox): convert files to typescript
