--- src/components/checkbox/index.d.ts
@@ -1,2 +1,0 @@
-export { default as CheckboxGroup } from "./checkbox-group";
-export { Checkbox } from "./checkbox";
