--- src/components/checkbox/index.js
@@ -1,2 +1,0 @@
-export { default as Checkbox } from "./checkbox.component";
-export { default as CheckboxGroup } from "./checkbox-group.component";
