+++ src/components/checkbox/index.ts
@@ -1,0 +1,4 @@
+export { default as CheckboxGroup } from "./checkbox-group.component";
+export type { CheckboxGroupProps } from "./checkbox-group.component";
+export { default as Checkbox } from "./checkbox.component";
+export type { CheckboxProps } from "./checkbox.component";
